backend/requirements.txt
alembic==1.18.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
logtail-python==0.3.4
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==25.0
pluggy==1.6.0
psycopg2==2.9.11
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlmodel==0.0.14
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.24.0
watchfiles==1.1.1
websockets==16.0
